+++ src/main/resources/db/migration/V007__create_product_table.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS tb_products (
     name VARCHAR(100) NOT NULL UNIQUE,
     measurement_unit VARCHAR(5) NOT NULL,
     unit_price DECIMAL(10,4) NOT NULL CHECK (unit_price > 0.0),
-    quantity_in_stock INT NOT NULL CHECK (quantity_in_stock >= 0),
+    quantity_in_stock SMALLINT NOT NULL CHECK (quantity_in_stock >= 0),
     category_id SMALLINT NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT now(),
     updated_at TIMESTAMP WITH TIME ZONE DEFAULT now(),
